+++ app/src/main/kotlin/com/pitchedapps/frost/activities/MainActivity.kt
@@ -50,7 +50,7 @@ import com.pitchedapps.frost.dbflow.loadFbCookie
 import com.pitchedapps.frost.dbflow.loadFbTabs
 import com.pitchedapps.frost.facebook.FbCookie
 import com.pitchedapps.frost.facebook.FbCookie.switchUser
-import com.pitchedapps.frost.facebook.FbTab
+import com.pitchedapps.frost.facebook.FbItem
 import com.pitchedapps.frost.facebook.PROFILE_PICTURE_URL
 import com.pitchedapps.frost.fragments.WebFragment
 import com.pitchedapps.frost.utils.*
@@ -60,6 +60,7 @@ import com.pitchedapps.frost.utils.iab.IS_FROST_PRO
 import com.pitchedapps.frost.views.BadgedIcon
 import com.pitchedapps.frost.views.FrostViewPager
 import com.pitchedapps.frost.web.SearchWebView
+import com.pitchedapps.frost.web.shouldLoadImages
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
 import io.reactivex.schedulers.Schedulers
@@ -160,14 +161,10 @@ class MainActivity : BaseActivity(), SearchWebView.SearchContract,
 //        }
         setFrostColors(toolbar, themeWindow = false, headers = arrayOf(tabs, appBar), backgrounds = arrayOf(viewPager))
         onCreateBilling()
-        if (Prefs.installDate < 1501454310304 && Showcase.intro)
-            materialDialogThemed {
-                title(R.string.intro_title)
-                content(R.string.intro_desc)
-                positiveText(R.string.kau_yes)
-                negativeText(R.string.kau_no)
-                onPositive { _, _ -> launchIntroActivity(cookies()) }
-            }
+        setNetworkObserver {
+            connectivity ->
+            shouldLoadImages = !connectivity.isRoaming
+        }
     }
 
     fun tabsForEachView(action: (position: Int, view: BadgedIcon) -> Unit) {
@@ -206,10 +203,10 @@ class MainActivity : BaseActivity(), SearchWebView.SearchContract,
                     tabsForEachView {
                         _, view ->
                         when (view.iicon) {
-                            FbTab.FEED.icon -> view.badgeText = feed
-                            FbTab.FRIENDS.icon -> view.badgeText = requests
-                            FbTab.MESSAGES.icon -> view.badgeText = messages
-                            FbTab.NOTIFICATIONS.icon -> view.badgeText = notifications
+                            FbItem.FEED.icon -> view.badgeText = feed
+                            FbItem.FRIENDS.icon -> view.badgeText = requests
+                            FbItem.MESSAGES.icon -> view.badgeText = messages
+                            FbItem.NOTIFICATIONS.icon -> view.badgeText = notifications
                         }
                     }
                 }
@@ -230,10 +227,10 @@ class MainActivity : BaseActivity(), SearchWebView.SearchContract,
             translucentStatusBar = false
             sliderBackgroundColor = navBg
             drawerHeader = accountHeader {
+                customViewRes = R.layout.material_drawer_header
                 textColor = Prefs.iconColor.toLong()
                 backgroundDrawable = ColorDrawable(navHeader)
                 selectionSecondLineShown = false
-                paddingBelow = false
                 cookies().forEach { (id, name) ->
                     profile(name = name ?: "") {
                         iconUrl = PROFILE_PICTURE_URL(id)
@@ -261,7 +258,7 @@ class MainActivity : BaseActivity(), SearchWebView.SearchContract,
                     identifier = -4L
                 }
                 onProfileChanged { _, profile, current ->
-                    if (current) launchWebOverlay(FbTab.PROFILE.url)
+                    if (current) launchWebOverlay(FbItem.PROFILE.url)
                     else when (profile.identifier) {
                         -2L -> {
                             val currentCookie = loadFbCookie(Prefs.userId)
@@ -295,25 +292,25 @@ class MainActivity : BaseActivity(), SearchWebView.SearchContract,
                 }
             }
             drawerHeader.setActiveProfile(Prefs.userId)
-            primaryFrostItem(FbTab.FEED_MOST_RECENT)
-            primaryFrostItem(FbTab.FEED_TOP_STORIES)
-            primaryFrostItem(FbTab.ACTIVITY_LOG)
+            primaryFrostItem(FbItem.FEED_MOST_RECENT)
+            primaryFrostItem(FbItem.FEED_TOP_STORIES)
+            primaryFrostItem(FbItem.ACTIVITY_LOG)
             divider()
-            primaryFrostItem(FbTab.PHOTOS)
-            primaryFrostItem(FbTab.GROUPS)
-            primaryFrostItem(FbTab.FRIENDS)
-            primaryFrostItem(FbTab.PAGES)
+            primaryFrostItem(FbItem.PHOTOS)
+            primaryFrostItem(FbItem.GROUPS)
+            primaryFrostItem(FbItem.FRIENDS)
+            primaryFrostItem(FbItem.PAGES)
             divider()
-            primaryFrostItem(FbTab.EVENTS)
-            primaryFrostItem(FbTab.BIRTHDAYS)
-            primaryFrostItem(FbTab.ON_THIS_DAY)
+            primaryFrostItem(FbItem.EVENTS)
+            primaryFrostItem(FbItem.BIRTHDAYS)
+            primaryFrostItem(FbItem.ON_THIS_DAY)
             divider()
-            primaryFrostItem(FbTab.NOTES)
-            primaryFrostItem(FbTab.SAVED)
+            primaryFrostItem(FbItem.NOTES)
+            primaryFrostItem(FbItem.SAVED)
         }
     }
 
-    fun Builder.primaryFrostItem(item: FbTab) = this.primaryItem(item.titleId) {
+    fun Builder.primaryFrostItem(item: FbItem) = this.primaryItem(item.titleId) {
         iicon = item.icon
         iconColor = Prefs.textColor.toLong()
         textColor = Prefs.textColor.toLong()
@@ -349,6 +346,7 @@ class MainActivity : BaseActivity(), SearchWebView.SearchContract,
     override fun searchOverlayDispose() {
         hiddenSearchView?.dispose()
         hiddenSearchView = null
+        searchView?.unBind { launchWebOverlay(FbItem.SEARCH.url); true }
         searchView = null
     }
 
@@ -369,10 +367,7 @@ class MainActivity : BaseActivity(), SearchWebView.SearchContract,
         if (Prefs.searchBar) {
             if (firstLoadFinished && hiddenSearchView == null) hiddenSearchView = SearchWebView(this, this)
             if (searchView == null) searchView = bindSearchView(menu, R.id.action_search, Prefs.iconColor) {
-                textObserver = {
-                    observable, _ ->
-                    observable.observeOn(AndroidSchedulers.mainThread()).subscribe { hiddenSearchView?.query(it) }
-                }
+                textCallback = { query, _ -> runOnUiThread { hiddenSearchView?.query(query) } }
                 foregroundColor = Prefs.textColor
                 backgroundColor = Prefs.bgColor.withMinAlpha(200)
                 openListener = { hiddenSearchView?.pauseLoad = false }
@@ -380,8 +375,8 @@ class MainActivity : BaseActivity(), SearchWebView.SearchContract,
                 onItemClick = { _, key, _, _ -> launchWebOverlay(key) }
             }
         } else {
-            searchOverlayDispose()
-            menu.findItem(R.id.action_search).setOnMenuItemClickListener { _ -> launchWebOverlay(FbTab.SEARCH.url); true }
+            if (searchView != null) searchOverlayDispose()
+            else menu.findItem(R.id.action_search).setOnMenuItemClickListener { _ -> launchWebOverlay(FbItem.SEARCH.url); true }
         }
         return true
     }
@@ -461,7 +456,7 @@ class MainActivity : BaseActivity(), SearchWebView.SearchContract,
     val currentFragment
         get() = supportFragmentManager.findFragmentByTag("android:switcher:${R.id.container}:${viewPager.currentItem}") as WebFragment
 
-    inner class SectionsPagerAdapter(fm: FragmentManager, val pages: List<FbTab>) : FragmentPagerAdapter(fm) {
+    inner class SectionsPagerAdapter(fm: FragmentManager, val pages: List<FbItem>) : FragmentPagerAdapter(fm) {
 
         override fun getItem(position: Int): Fragment {
             val fragment = WebFragment(pages[position], position)