+++ app/src/main/kotlin/com/pitchedapps/frost/services/FrostNotifications.kt
@@ -21,7 +21,6 @@ import com.pitchedapps.frost.BuildConfig
 import com.pitchedapps.frost.R
 import com.pitchedapps.frost.activities.FrostWebActivity
 import com.pitchedapps.frost.dbflow.CookieModel
-import com.pitchedapps.frost.dbflow.fetchUsername
 import com.pitchedapps.frost.facebook.formattedFbUrl
 import com.pitchedapps.frost.utils.ARG_USER_ID
 import com.pitchedapps.frost.utils.L
@@ -31,6 +30,12 @@ import org.jetbrains.anko.runOnUiThread
 
 /**
  * Created by Allan Wang on 2017-07-08.
+ *
+ * Logic for build notifications, scheduling notifications, and showing notifications
+ */
+
+/**
+ * Wrap the default builder with our icon and accent color
  */
 val Context.frostNotification: NotificationCompat.Builder
     get() = NotificationCompat.Builder(this, BuildConfig.APPLICATION_ID).apply {
@@ -39,6 +44,9 @@ val Context.frostNotification: NotificationCompat.Builder
         color = color(R.color.frost_notification_accent)
     }
 
+/**
+ * Assign global changes to the notification after it is built
+ */
 @Suppress("DEPRECATION")
         //The update feature is for Android O and seems to still be in beta
 fun Notification.frostConfig() = apply {
@@ -54,6 +62,7 @@ val NotificationCompat.Builder.withBigText: NotificationCompat.BigTextStyle
  * Created by Allan Wang on 2017-07-08.
  *
  * Custom target to set the content view and update a given notification
+ * 40dp is the size of the right avatar
  */
 class FrostNotificationTarget(val context: Context,
                               val notifId: Int,
@@ -67,6 +76,9 @@ class FrostNotificationTarget(val context: Context,
     }
 }
 
+internal const val FROST_NOTIFICATION_GROUP = "frost"
+internal const val FROST_MESSAGE_NOTIFICATION_GROUP = "frost_im"
+
 /**
  * Notification data holder
  */
@@ -77,39 +89,35 @@ data class NotificationContent(val data: CookieModel,
                                val text: String,
                                val timestamp: Long,
                                val profileUrl: String) {
-    fun createNotification(context: Context, verifiedUser: Boolean = false) {
-        //in case we haven't found the name, we will try one more time before passing the notification
-        if (!verifiedUser && data.name?.isBlank() ?: true) {
-            data.fetchUsername {
-                data.name = it
-                createNotification(context, true)
-            }
-        } else {
-            val intent = Intent(context, FrostWebActivity::class.java)
-            intent.data = Uri.parse(href.formattedFbUrl)
-            intent.putExtra(ARG_USER_ID, data.id)
-            val group = "frost_${data.id}"
-            val pendingIntent = PendingIntent.getActivity(context, 0, intent, 0)
-            val notifBuilder = context.frostNotification
-                    .setContentTitle(title ?: context.string(R.string.frost_name))
-                    .setContentText(text)
-                    .setContentIntent(pendingIntent)
-                    .setCategory(Notification.CATEGORY_SOCIAL)
-                    .setSubText(data.name)
-                    .setGroup(group)
+    fun createNotification(context: Context) = createNotification(context, FROST_NOTIFICATION_GROUP)
 
-            if (timestamp != -1L) notifBuilder.setWhen(timestamp * 1000)
-            L.v("Notif load $this")
-            NotificationManagerCompat.from(context).notify(group, notifId, notifBuilder.withBigText.build().frostConfig())
+    fun createMessageNotification(context: Context) = createNotification(context, FROST_MESSAGE_NOTIFICATION_GROUP)
 
-            if (profileUrl.isNotBlank()) {
-                context.runOnUiThread {
-                    Glide.with(context)
-                            .asBitmap()
-                            .load(profileUrl)
-                            .withRoundIcon()
-                            .into(FrostNotificationTarget(context, notifId, group, notifBuilder))
-                }
+    private fun createNotification(context: Context, groupPrefix: String) {
+        val intent = Intent(context, FrostWebActivity::class.java)
+        intent.data = Uri.parse(href.formattedFbUrl)
+        intent.putExtra(ARG_USER_ID, data.id)
+        val group = "${groupPrefix}_${data.id}"
+        val pendingIntent = PendingIntent.getActivity(context, 0, intent, 0)
+        val notifBuilder = context.frostNotification
+                .setContentTitle(title ?: context.string(R.string.frost_name))
+                .setContentText(text)
+                .setContentIntent(pendingIntent)
+                .setCategory(Notification.CATEGORY_SOCIAL)
+                .setSubText(data.name)
+                .setGroup(group)
+
+        if (timestamp != -1L) notifBuilder.setWhen(timestamp * 1000)
+        L.v("Notif load", this.toString())
+        NotificationManagerCompat.from(context).notify(group, notifId, notifBuilder.withBigText.build().frostConfig())
+
+        if (profileUrl.isNotBlank()) {
+            context.runOnUiThread {
+                Glide.with(context)
+                        .asBitmap()
+                        .load(profileUrl)
+                        .withRoundIcon()
+                        .into(FrostNotificationTarget(context, notifId, group, notifBuilder))
             }
         }
     }