+++ app/src/main/kotlin/com/pitchedapps/frost/services/NotificationService.kt
@@ -1,26 +1,33 @@
 package com.pitchedapps.frost.services
 
+import android.app.Notification
+import android.app.PendingIntent
 import android.app.job.JobParameters
 import android.app.job.JobService
 import android.content.Context
+import android.content.Intent
+import android.net.Uri
 import android.support.v4.app.NotificationManagerCompat
 import ca.allanwang.kau.utils.string
 import com.pitchedapps.frost.BuildConfig
 import com.pitchedapps.frost.R
+import com.pitchedapps.frost.activities.FrostWebActivity
 import com.pitchedapps.frost.dbflow.CookieModel
 import com.pitchedapps.frost.dbflow.lastNotificationTime
 import com.pitchedapps.frost.dbflow.loadFbCookie
 import com.pitchedapps.frost.dbflow.loadFbCookiesSync
 import com.pitchedapps.frost.facebook.FACEBOOK_COM
-import com.pitchedapps.frost.facebook.FbTab
+import com.pitchedapps.frost.facebook.FbItem
 import com.pitchedapps.frost.facebook.USER_AGENT_BASIC
 import com.pitchedapps.frost.facebook.formattedFbUrl
+import com.pitchedapps.frost.injectors.JsAssets
+import com.pitchedapps.frost.utils.ARG_USER_ID
 import com.pitchedapps.frost.utils.L
 import com.pitchedapps.frost.utils.Prefs
 import com.pitchedapps.frost.utils.frostAnswersCustom
-import com.pitchedapps.frost.web.MessageWebView
+import com.pitchedapps.frost.web.launchHeadlessHtmlExtractor
+import io.reactivex.schedulers.Schedulers
 import org.jetbrains.anko.doAsync
-import org.jetbrains.anko.uiThread
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Element
 import java.util.concurrent.Future
@@ -30,6 +37,9 @@ import java.util.concurrent.Future
  *
  * Service to manage notifications
  * Will periodically check through all accounts in the db and send notifications when appropriate
+ *
+ * Note that general notifications are parsed directly with Jsoup,
+ * but instant messages are done so with a headless webview as it is generated from JS
  */
 class NotificationService : JobService() {
 
@@ -58,7 +68,7 @@ class NotificationService : JobService() {
 
     fun finish(params: JobParameters?) {
         val time = System.currentTimeMillis() - startTime
-        L.d("Notification service has finished in $time ms")
+        L.i("Notification service has finished in $time ms")
         frostAnswersCustom("NotificationTime",
                 "Type" to "Service",
                 "IM Included" to Prefs.notificationsInstantMessages,
@@ -68,8 +78,8 @@ class NotificationService : JobService() {
         future = null
     }
 
-
     override fun onStartJob(params: JobParameters?): Boolean {
+        L.i("Fetching notifications")
         future = doAsync {
             if (Prefs.notificationAllAccounts) {
                 val cookies = loadFbCookiesSync()
@@ -83,9 +93,15 @@ class NotificationService : JobService() {
             L.d("Finished main notifications")
             if (Prefs.notificationsInstantMessages) {
                 val currentCookie = loadFbCookie(Prefs.userId)
-                if (currentCookie != null)
-                    uiThread { MessageWebView(this@NotificationService, params, currentCookie) }
-            } else finish(params)
+                if (currentCookie != null) {
+                    fetchMessageNotifications(currentCookie) {
+                        L.i("Notif IM fetching finished ${if (it) "succesfully" else "unsuccessfully"}")
+                        finish(params)
+                    }
+                    return@doAsync
+                }
+            }
+            finish(params)
         }
         return true
     }
@@ -95,13 +111,21 @@ class NotificationService : JobService() {
         return null
     }
 
+    /*
+     * ----------------------------------------------------------------
+     * General notification logic.
+     * Fetch notifications -> Filter new ones -> Parse notifications ->
+     * Show notifications -> Show group notification
+     * ----------------------------------------------------------------
+     */
+
     fun fetchGeneralNotifications(data: CookieModel) {
-        L.i("Notif fetch for $data")
-        val doc = Jsoup.connect(FbTab.NOTIFICATIONS.url).cookie(FACEBOOK_COM, data.cookie).userAgent(USER_AGENT_BASIC).get()
+        L.d("Notif fetch", data.toString())
+        val doc = Jsoup.connect(FbItem.NOTIFICATIONS.url).cookie(FACEBOOK_COM, data.cookie).userAgent(USER_AGENT_BASIC).get()
         //aclb for unread, acw for read
         val unreadNotifications = (doc.getElementById("notifications_list") ?: return L.eThrow("Notification list not found")).getElementsByClass("aclb")
         var notifCount = 0
-//        val prevLatestEpoch = 1498931565L // for testing
+        //val prevLatestEpoch = 1498931565L // for testing
         val prevNotifTime = lastNotificationTime(data.id)
         val prevLatestEpoch = prevNotifTime.epoch
         L.v("Notif Prev Latest Epoch $prevLatestEpoch")
@@ -122,7 +146,6 @@ class NotificationService : JobService() {
         summaryNotification(data.id, notifCount)
     }
 
-
     fun parseNotification(data: CookieModel, element: Element): NotificationContent? {
         val a = element.getElementsByTag("a").first() ?: return logNotif("IM No a tag")
         val abbr = element.getElementsByTag("abbr")
@@ -134,13 +157,36 @@ class NotificationService : JobService() {
         if (Prefs.notificationKeywords.any { text.contains(it, ignoreCase = true) }) return null //notification filtered out
         //fetch profpic
         val p = element.select("i.img[style*=url]")
-        val pUrl = profMatcher.find(p.attr("style"))?.groups?.get(1)?.value ?: ""
+        val pUrl = profMatcher.find(p.attr("style"))?.groups?.get(1)?.value?.formattedFbUrl ?: ""
         return NotificationContent(data, notifId.toInt(), a.attr("href"), null, text, epoch, pUrl)
     }
 
-    fun fetchMessageNotifications(data: CookieModel, content: String) {
-        L.i("Notif IM fetch for $data")
-        val doc = Jsoup.parseBodyFragment(content)
+    fun summaryNotification(userId: Long, count: Int)
+            = summaryNotification(userId, count, R.string.notifications, FbItem.NOTIFICATIONS.url, FROST_NOTIFICATION_GROUP)
+
+    /*
+     * ----------------------------------------------------------------
+     * Instant message notification logic.
+     * Fetch notifications -> Filter new ones -> Parse notifications ->
+     * Show notifications -> Show group notification
+     * ----------------------------------------------------------------
+     */
+
+    inline fun fetchMessageNotifications(data: CookieModel, crossinline callback: (success: Boolean) -> Unit) {
+        launchHeadlessHtmlExtractor(FbItem.MESSAGES.url, JsAssets.NOTIF_MSG) {
+            it.observeOn(Schedulers.newThread()).subscribe {
+                (html, errorRes) ->
+                L.d("Notf IM html received")
+                if (errorRes != -1) return@subscribe callback(false)
+                fetchMessageNotifications(data, html)
+                callback(true)
+            }
+        }
+    }
+
+    fun fetchMessageNotifications(data: CookieModel, html: String) {
+        L.d("Notif IM fetch", data.toString())
+        val doc = Jsoup.parseBodyFragment(html)
         val unreadNotifications = (doc.getElementById("threadlist_rows") ?: return L.eThrow("Notification messages not found")).getElementsByClass("aclb")
         var notifCount = 0
         val prevNotifTime = lastNotificationTime(data.id)
@@ -152,7 +198,7 @@ class NotificationService : JobService() {
             val notif = parseMessageNotification(data, elem) ?: return@unread
             L.v("Notif im timestamp ${notif.timestamp}")
             if (notif.timestamp <= prevLatestEpoch) return@unread
-            notif.createNotification(this@NotificationService)
+            notif.createMessageNotification(this@NotificationService)
             if (notif.timestamp > newLatestEpoch)
                 newLatestEpoch = notif.timestamp
             notifCount++
@@ -160,7 +206,7 @@ class NotificationService : JobService() {
         if (newLatestEpoch != prevLatestEpoch) prevNotifTime.copy(epochIm = newLatestEpoch).save()
         L.d("Notif new latest im epoch ${lastNotificationTime(data.id).epochIm}")
         frostAnswersCustom("Notifications", "Type" to "Message", "Count" to notifCount)
-        summaryNotification(data.id, notifCount)
+        summaryMessageNotification(data.id, notifCount)
     }
 
     fun parseMessageNotification(data: CookieModel, element: Element): NotificationContent? {
@@ -174,10 +220,13 @@ class NotificationService : JobService() {
         if (Prefs.notificationKeywords.any { text.contains(it, ignoreCase = true) }) return null //notification filtered out
         //fetch convo pic
         val p = element.select("i.img[style*=url]")
-        val pUrl = profMatcher.find(p.attr("style"))?.groups?.get(1)?.value ?: ""
-        L.v("url ${a.attr("href")}")
-        return NotificationContent(data, notifId.toInt(), a.attr("href"), a.text(), text, epoch, pUrl.formattedFbUrl)
+        val pUrl = profMatcher.find(p.attr("style"))?.groups?.get(1)?.value?.formattedFbUrl ?: ""
+        L.v("url", a.attr("href"))
+        return NotificationContent(data, notifId.toInt(), a.attr("href"), a.text(), text, epoch, pUrl)
     }
+
+    fun summaryMessageNotification(userId: Long, count: Int)
+            = summaryNotification(userId, count, R.string.messages, FbItem.MESSAGES.url, FROST_MESSAGE_NOTIFICATION_GROUP)
 
     private fun Context.debugNotification(text: String) {
         if (!BuildConfig.DEBUG) return
@@ -187,15 +236,21 @@ class NotificationService : JobService() {
         NotificationManagerCompat.from(this).notify(999, notifBuilder.build().frostConfig())
     }
 
-    fun summaryNotification(userId: Long, count: Int) {
+    private fun summaryNotification(userId: Long, count: Int, contentRes: Int, pendingUrl: String, groupPrefix: String) {
         if (count <= 1) return
+        val intent = Intent(this, FrostWebActivity::class.java)
+        intent.data = Uri.parse(pendingUrl)
+        intent.putExtra(ARG_USER_ID, userId)
+        val pendingIntent = PendingIntent.getActivity(this, 0, intent, 0)
         val notifBuilder = frostNotification
                 .setContentTitle(string(R.string.frost_name))
-                .setContentText("$count notifications")
-                .setGroup("frost_$userId")
+                .setContentText("$count ${string(contentRes)}")
+                .setGroup("${groupPrefix}_$userId")
                 .setGroupSummary(true)
+                .setContentIntent(pendingIntent)
+                .setCategory(Notification.CATEGORY_SOCIAL)
 
-        NotificationManagerCompat.from(this).notify("frost_$userId", userId.toInt(), notifBuilder.build().frostConfig())
+        NotificationManagerCompat.from(this).notify("${groupPrefix}_$userId", userId.toInt(), notifBuilder.build().frostConfig())
     }
 
 }