+++ app/src/main/kotlin/com/pitchedapps/frost/web/FrostRequestInterceptor.kt
@@ -5,6 +5,7 @@ import android.webkit.WebResourceResponse
 import android.webkit.WebView
 import ca.allanwang.kau.utils.use
 import com.pitchedapps.frost.utils.L
+import com.pitchedapps.frost.utils.Prefs
 import okhttp3.HttpUrl
 import java.io.ByteArrayInputStream
 
@@ -15,17 +16,17 @@ import java.io.ByteArrayInputStream
  * Handler to decide when a request should be done by us
  * This is the crux of Frost's optimizations for the web browser
  */
-val blankResource: WebResourceResponse by lazy { WebResourceResponse("text/plain", "utf-8", ByteArrayInputStream("".toByteArray())) }
+private val blankResource: WebResourceResponse by lazy { WebResourceResponse("text/plain", "utf-8", ByteArrayInputStream("".toByteArray())) }
 
 //these hosts will redirect to a blank resource
-val blacklistHost: Set<String> by lazy {
+private val blacklistHost: Set<String> by lazy {
     setOf(
             "edge-chat.facebook.com"
     )
 }
 
 //these hosts will return null and skip logging
-val whitelistHost: Set<String> by lazy {
+private val whitelistHost: Set<String> by lazy {
     setOf(
             "static.xx.fbcdn.net",
             "m.facebook.com",
@@ -35,13 +36,13 @@ val whitelistHost: Set<String> by lazy {
 
 //these hosts will skip ad inspection
 //this list does not have to include anything from the two above
-val adWhitelistHost: Set<String> by lazy {
+private val adWhitelistHost: Set<String> by lazy {
     setOf(
             "scontent-sea1-1.xx.fbcdn.net"
     )
 }
 
-var adblock: Set<String>? = null
+private var adblock: Set<String>? = null
 
 fun shouldFrostInterceptRequest(view: WebView, request: WebResourceRequest): WebResourceResponse? {
     val httpUrl = HttpUrl.parse(request.url?.toString() ?: return null) ?: return null
@@ -53,7 +54,8 @@ fun shouldFrostInterceptRequest(view: WebView, request: WebResourceRequest): Web
         if (adblock == null) adblock = view.context.assets.open("adblock.txt").bufferedReader().use { it.readLines().toSet() }
         if (adblock?.any { url.contains(it) } ?: false) return blankResource
     }
-    L.v("Intercept Request ${host} ${url}")
+    if (!shouldLoadImages && !Prefs.loadMediaOnMeteredNetwork && request.isMedia) return blankResource
+    L.v("Intercept Request", "$host $url")
     return null
 }
 
@@ -64,16 +66,25 @@ fun WebResourceRequest.query(action: (url: String) -> Boolean): Boolean {
     return action(url?.path ?: return false)
 }
 
+val WebResourceRequest.isImage: Boolean
+    get() = query { it.contains(".jpg") || it.contains(".png") }
+
+val WebResourceRequest.isMedia: Boolean
+    get() = query { it.contains(".jpg") || it.contains(".png") || it.contains("video") }
+
 /**
  * Generic filter passthrough
  * If Resource is already nonnull, pass it, otherwise check if filter is met and override the response accordingly
  */
-fun WebResourceResponse?.filter(request: WebResourceRequest, filter: (url: String) -> Boolean): WebResourceResponse?
-        = this ?: if (request.query { filter(it) }) blankResource else null
+fun WebResourceResponse?.filter(request: WebResourceRequest, filter: (url: String) -> Boolean)
+        = filter(request.query { filter(it) })
+
+fun WebResourceResponse?.filter(filter: Boolean): WebResourceResponse?
+        = this ?: if (filter) blankResource else null
 
 fun WebResourceResponse?.filterCss(request: WebResourceRequest): WebResourceResponse?
         = filter(request) { it.endsWith(".css") }
 
 fun WebResourceResponse?.filterImage(request: WebResourceRequest): WebResourceResponse?
-        = filter(request) { it.contains(".jpg") || it.contains(".png") }
+        = filter(request.isImage)