+++ app/src/main/kotlin/com/pitchedapps/frost/web/SearchWebView.kt
@@ -6,7 +6,7 @@ import android.webkit.JavascriptInterface
 import android.webkit.WebView
 import ca.allanwang.kau.searchview.SearchItem
 import ca.allanwang.kau.utils.gone
-import com.pitchedapps.frost.facebook.FbTab
+import com.pitchedapps.frost.facebook.FbItem
 import com.pitchedapps.frost.facebook.USER_AGENT_BASIC
 import com.pitchedapps.frost.injectors.JsAssets
 import com.pitchedapps.frost.injectors.JsBuilder
@@ -27,7 +27,7 @@ import java.util.concurrent.TimeUnit
  */
 class SearchWebView(context: Context, val contract: SearchContract) : WebView(context) {
 
-    val searchSubject = PublishSubject.create<String>()
+    val searchSubject = PublishSubject.create<String>()!!
 
     init {
         gone()
@@ -39,11 +39,11 @@ class SearchWebView(context: Context, val contract: SearchContract) : WebView(co
      * Contains the last item's href (search more) as well as the number of items found
      * This holder is synchronized
      */
-    var previousResult: Pair<String?, Int> = Pair(null, 0)
+    var previousResult: Pair<String, Int> = Pair("", 0)
 
     fun saveResultFrame(result: List<Pair<List<String>, String>>) {
         synchronized(previousResult) {
-            previousResult = Pair(result.lastOrNull()?.second, result.size)
+            previousResult = Pair(result.last().second, result.size)
         }
     }
 
@@ -56,17 +56,22 @@ class SearchWebView(context: Context, val contract: SearchContract) : WebView(co
         addJavascriptInterface(SearchJSI(), "Frost")
         searchSubject.debounce(300, TimeUnit.MILLISECONDS).subscribeOn(Schedulers.newThread())
                 .map {
-                    Jsoup.parse(it).select("a:not([rel*='keywords(']):not([href=#])[rel]").map {
+                    val doc = Jsoup.parse(it)
+                    L.d(doc.getElementById("main-search_input")?.html())
+                    val searchQuery = doc.getElementById("main-search-input")?.text() ?: "Null input"
+                    L.d("Search query", searchQuery)
+                    doc.select("a:not([rel*='keywords(']):not([href=#])[rel]").map {
                         element ->
                         //split text into separate items
-                        L.v("Search element ${element.attr("href")}")
-                        val texts = element.select("div").map { (it.text()) }.filter { it.isNotBlank() }
+                        L.v("Search element", element.attr("href"))
+                        val texts = element.select("div").map { it.text() }.filter { !it.isNullOrBlank() }
                         val pair = Pair(texts, element.attr("href"))
-                        L.v("Search element potential $pair")
+                        L.v("Search element potential", pair.toString())
                         pair
                     }.filter { it.first.isNotEmpty() }
                 }
-                .filter { content -> Pair(content.lastOrNull()?.second, content.size) != previousResult }
+                .filter { it.isNotEmpty() }
+                .filter { Pair(it.last().second, it.size) != previousResult }
                 .subscribe {
                     content: List<Pair<List<String>, String>> ->
                     saveResultFrame(content)
@@ -90,7 +95,7 @@ class SearchWebView(context: Context, val contract: SearchContract) : WebView(co
         }
 
     override fun reload() {
-        super.loadUrl(FbTab.SEARCH.url)
+        super.loadUrl(FbItem.SEARCH.url)
     }
 
     /**
@@ -104,7 +109,8 @@ class SearchWebView(context: Context, val contract: SearchContract) : WebView(co
 
     inner class SearchJSI {
         @JavascriptInterface
-        fun handleHtml(html: String) {
+        fun handleHtml(html: String?) {
+            html ?: return
             L.d("Search received response ${contract.isSearchOpened}")
             if (!contract.isSearchOpened) pauseLoad = true
             searchSubject.onNext(html)
@@ -117,10 +123,13 @@ class SearchWebView(context: Context, val contract: SearchContract) : WebView(co
                     L.d("Search loaded successfully")
                 }
                 1 -> { //something is not found in the search view; this is effectively useless
-                    L.eThrow("Search subject error; reverting to full overlay")
+                    L.e("Search subject error; reverting to full overlay")
                     Prefs.searchBar = false
                     searchSubject.onComplete()
                     contract.searchOverlayDispose()
+                }
+                2 -> {
+                    L.v("Search emission received")
                 }
             }
         }