+++ app/src/test/kotlin/com/pitchedapps/frost/MiscTest.kt
@@ -2,6 +2,7 @@ package com.pitchedapps.frost
 
 import com.pitchedapps.frost.injectors.CssHider
 import org.junit.Test
+import kotlin.test.assertEquals
 
 /**
  * Created by Allan Wang on 2017-06-14.
@@ -9,7 +10,12 @@ import org.junit.Test
 class MiscTest {
 
     @Test
-    fun asdf() {
+    fun headerFunction() {
         print(CssHider.HEADER.injector.function)
+    }
+
+    @Test
+    fun nullPair() {
+        assertEquals(Pair<String?, Int>(null, 2), Pair<String?, Int>(null, 2))
     }
 }